+++ lib/news/view_news.dart
@@ -135,7 +135,10 @@ class _ViewNewsState extends State<ViewNews> {
                     showPathInfo: true,
                   ),
                   // Display markdown content
-                  displayPostText(_newsPost.content!.newsContent),
+                  displayPostText(
+                    article: _newsPost.content!.newsContent,
+                    imageCache: _newsPost.cachedImages,
+                  ),
                 ],
               ),
             ),
@@ -164,6 +167,7 @@ class _ViewNewsState extends State<ViewNews> {
                             newsUrl: _newsPost.newsUrl,
                             newsOwner: _newsPost.newsOwner,
                             isExternal: _newsPost.isExternalRes,
+                            newsContent: _newsPost.content,
                             backPage: ViewNews(
                               newsPost: _newsPost,
                               scaffoldController: _scaffoldController,